基本信息修改

+++ boatBoot/src/main/java/com/wisely/service/BaseMetaBackingService.java
@@ -10,8 +10,10 @@ import org.springframework.stereotype.Service;
 
 import com.wisely.dao.BaseMetaBackingDao;
 import com.wisely.dao.PhotoDao;
+import com.wisely.domain.common.BaseMetaSample;
 import com.wisely.domain.common.Photo;
 import com.wisely.domain.small.BaseMetaBacking;
+import com.wisely.domainVO.ResultVO;
 import com.wisely.util.Toolkit;
 
 /**
@@ -26,6 +28,8 @@ public class BaseMetaBackingService{
 	@Autowired
 	private PhotoDao photoDao;
 	
+	@Autowired
+	private UpdateColumnService updateColumnService;
 	
 	public BaseMetaBacking findByID(String pk){
 		return dao.findOne(pk);
@@ -73,6 +77,7 @@ public class BaseMetaBackingService{
 				result = dao.save(entity); 
 				if(Toolkit.notEmpty(result)){
 					updatePhoto(result.getPk(), photos);
+					updateColumnService.updateValue("smallmetadata", "backgroundtype", result.getName(), "backing_pk", result.getPk());
 				}
 			}
 		}
@@ -93,11 +98,23 @@ public class BaseMetaBackingService{
 	 * 删除实体
 	 * @param pk
 	 */
-	public void deleteEntity(String pk){
+	public ResultVO deleteEntity(String pk){
+		ResultVO re = new ResultVO(false);
+		int count = updateColumnService.queryByColumn("smallmetadata", "backing_pk", pk);
 		if(Toolkit.notEmpty(pk)){
-			photoDao.modifyDeleted(pk,1);
-			dao.delete(pk);
+			if(count == 0){
+				photoDao.modifyDeleted(pk,1);
+				dao.delete(pk);
+				re.setSuccess(true);
+			}else{
+				re.setMessage("该数据被引用，不能直接删除");
+			}
+		}else{
+			re.setMessage("参数不能为空");
 		}
+		return re;
+		
+		
 	}
 	
 	public List<BaseMetaBacking> findAll() {
@@ -110,6 +127,10 @@ public class BaseMetaBackingService{
 			return result.get(0);
 		}
 		return null;
+	}
+	
+	public int queryOtherNameCount(BaseMetaBacking entity){
+		return dao.findOtherName(entity.getName(), entity.getPk());
 	}
