+++ boatBoot/src/main/java/com/wisely/service/scale/LayingSchemeService.java
@@ -12,7 +12,8 @@ import com.wisely.dao.PhotoDao;
 import com.wisely.dao.scale.LayingSchemeDao;
 import com.wisely.domain.common.Photo;
 import com.wisely.domain.scale.LayingSchemePO;
-import com.wisely.domain.scale.TestConditionPO;
+import com.wisely.domainVO.ResultVO;
+import com.wisely.service.UpdateColumnService;
 import com.wisely.util.Toolkit;
 
 @Service
@@ -20,9 +21,11 @@ public class LayingSchemeService {
 
 	@Autowired
 	private LayingSchemeDao dao;
-
 	@Autowired
 	private PhotoDao photoDao;
+	@Autowired
+	private UpdateColumnService updateColumnService;
+	
 	/**
 	 * 保存实体
 	 * @param entity
@@ -65,6 +68,7 @@ public class LayingSchemeService {
 				result = dao.save(entity); 
 				if(Toolkit.notEmpty(result)){
 					updatePhoto(result.getPk(), photos);
+					updateColumnService.updateValue("scale_mata", "laying_scheme_name", result.getName(), "laying_scheme_pk", result.getPk());
 				}
 			}
 		}
@@ -85,11 +89,21 @@ public class LayingSchemeService {
 	 * 删除实体
 	 * @param pk
 	 */
-	public void deleteEntity(String pk){
+	public ResultVO deleteEntity(String pk){
+		ResultVO re = new ResultVO(false);
+		int count = updateColumnService.queryByColumn("scale_mata", "laying_scheme_pk", pk);
 		if(Toolkit.notEmpty(pk)){
-			photoDao.modifyDeleted(pk,1);
-			dao.delete(pk);
+			if(count == 0){
+				photoDao.modifyDeleted(pk,1);
+				dao.delete(pk);
+				re.setSuccess(true);
+			}else{
+				re.setMessage("该数据被引用，不能直接删除");
+			}
+		}else{
+			re.setMessage("参数不能为空");
 		}
+		return re;
 	}
 	
 	
@@ -104,4 +118,9 @@ public class LayingSchemeService {
 	public LayingSchemePO getByName(String name) {
 		return dao.findByName(name);
 	}
+	public int queryOtherNameCount(LayingSchemePO entity){
+		return dao.findOtherName(entity.getName(), entity.getPk());
+	}
+	
+	
 }
